refactor for testing, add unit tests
