+++ mellanox_os/src/__init__.py
@@ -1,3 +1,2 @@
-# -*- coding: utf-8 -*-
 from pkgutil import extend_path
 __path__ = extend_path(__path__, __name__)
