+++ mellanox_os/src/driver.py
@@ -6,10 +6,12 @@ import re
 
 ##########################################
 # Add bugfix that was not released yet
-# This must be above import statements for cloudshell.networking.apply_connectivity
+# This must be above import statements for cloudshell.networking.*
 # https://github.com/QualiSystems/cloudshell-networking/blob/dev/cloudshell/networking/apply_connectivity/models/connectivity_request.py
 import cloudshell
-fn = os.path.join(os.path.dirname(cloudshell.__file__), 'networking', 'apply_connectivity', 'models', 'connectivity_request.py')
+
+fn = os.path.join(os.path.dirname(cloudshell.__file__), 'networking', 'apply_connectivity', 'models',
+                  'connectivity_request.py')
 with open(fn, 'r') as f:
     s = f.read()
 if "con_params = ConnectionParams()" in s and "dictionary['vlanId']" not in s:
@@ -22,147 +24,136 @@ if "con_params = ConnectionParams()" in s and "dictionary['vlanId']" not in s:
 
 from cloudshell.api.cloudshell_api import CloudShellAPISession
 from cloudshell.shell.core.resource_driver_interface import ResourceDriverInterface
-from cloudshell.shell.core.context import InitCommandContext, ResourceCommandContext, AutoLoadResource, AutoLoadAttribute, AutoLoadDetails
+from cloudshell.shell.core.context import AutoLoadCommandContext, InitCommandContext, ResourceCommandContext, \
+    AutoLoadResource, AutoLoadAttribute, AutoLoadDetails
+from cloudshell.shell.core.context_utils import get_attribute_by_name
 
-from cloudshell.networking.apply_connectivity.apply_connectivity_operation import apply_connectivity_changes, ConnectivityActionRequest
-from cloudshell.networking.apply_connectivity.models.connectivity_result import ConnectivitySuccessResponse, ConnectivityErrorResponse
+from cloudshell.networking.apply_connectivity.apply_connectivity_operation import apply_connectivity_changes, \
+    ConnectivityActionRequest
+from cloudshell.networking.apply_connectivity.models.connectivity_result import ConnectivitySuccessResponse, \
+    ConnectivityErrorResponse
 
 from cloudshell.core.logger.qs_logger import get_qs_logger
 
-class MellanoxOsDriver (ResourceDriverInterface):
-    def _log(self, context, message):
-        try:
-            try:
-                resid = context.reservation.reservation_id
-            except:
-                resid = 'out-of-reservation'
-            try:
-                resourcename = context.resource.fullname
-            except:
-                resourcename = 'no-resource'
-            logger = get_qs_logger(resid, 'Mellanox-OS-L2', resourcename)
-            logger.info(message)
-        except Exception as e:
-            try:
-                with open(r'c:\programdata\qualisystems\mellanox-os.log', 'a') as f:
-                    f.write(time.strftime('%Y-%m-%d %H:%M:%S', time.localtime()) + ' qs_logger failed: ' + str(e)+'\r\n')
-                    f.write(time.strftime('%Y-%m-%d %H:%M:%S', time.localtime()) + ' (QS LOGGER NOT WORKING): ' + message+'\r\n')
-            except:
-                pass
 
-    def _ssh_disconnect(self, context, ssh, channel):
-        self._log(context, 'disconnnect')
-        if self.fakedata:
-            return
-        ssh.close()
+class SSHManager:
+    def __init__(self, logger, host, port, username, password, prompt_regex):
+        self.logger = logger
+        self.logger.info('connect %s %d %s %s %s' % (host, port, username, password, prompt_regex))
+        self.ssh = paramiko.SSHClient()
+        self.ssh.load_system_host_keys()
+        self.ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
 
-    def _ssh_connect(self, context, host, port, username, password, prompt_regex):
-        self._log(context, 'connect %s %d %s %s %s' % (host, port, username, password, prompt_regex))
-        if self.fakedata:
-            return
-        ssh = paramiko.SSHClient()
-        ssh.load_system_host_keys()
-        ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-        ssh.connect(host,
-                    port=port,
-                    username=username,
-                    password=password,
-                    look_for_keys=True)
-        channel = ssh.invoke_shell(term='console', width=300, height=100000)
-        return ssh, channel, self._ssh_read(context, ssh, channel, prompt_regex)  # eat banner
+        self.ssh.connect(host,
+                         port=port,
+                         username=username,
+                         password=password,
+                         look_for_keys=True)
 
-    def _ssh_write(self, context, ssh, channel, command):
-        self._log(context, 'sending: <<<' + command + '>>>')
-        if self.fakedata:
-            print command
-            return
-        channel.send(command)
-        self._log(context, 'send complete')
+        self.channel = self.ssh.invoke_shell(term='console', width=300, height=100000)
+        self._read(prompt_regex)  # eat banner
 
-    def _ssh_read(self, context, ssh, channel, prompt_regex):
-        if self.fakedata:
-            return
-        rv = ''
-        self._log(context, 'read...')
+    def disconnect(self):
+        self.logger.info('disconnnect')
+        self.ssh.close()
+
+    def _write(self, command):
+        self.logger.info('sending: <<<' + command + '>>>')
+        self.channel.send(command)
+        self.logger.info('send complete')
+
+    def _read(self, prompt_regex):
+        buf = ''
+        self.logger.info('read...')
         while True:
-            self._log(context, 'recv')
-            r = channel.recv(2048)
-            self._log(context, 'recv returned: <<<' + str(r) + '>>>')
+            self.logger.info('recv')
+            r = self.channel.recv(2048)
+            self.logger.info('recv returned: <<<' + str(r) + '>>>')
             if r:
-                rv += r
-            if rv:
-                t = rv
-                t = re.sub(r'(\x9b|\x1b)[[?;0-9]*[a-zA-Z]', '', t)
-                t = re.sub(r'(\x9b|\x1b)[>=]', '', t)
-                t = re.sub('.\b', '', t) # not r''
+                buf += r
+
+            # if isinstance(r, str):
+            #     buf += r
+            # elif r is None:
+            #     continue
+            # else:
+            #     return ''
+
+            if buf:
+                cleanbuf = buf
+                cleanbuf = re.sub(r'(\x9b|\x1b)[[?;0-9]*[a-zA-Z]', '', cleanbuf)
+                cleanbuf = re.sub(r'(\x9b|\x1b)[>=]', '', cleanbuf)
+                cleanbuf = re.sub('.\b', '', cleanbuf)  # not r''
             else:
-                t = ''
-            if not r or len(re.findall(prompt_regex, t)) > 0:
-                rv = t
-                if rv:
-                    rv = rv.replace('\r', '\n')
-                self._log(context, '\n\nread complete: <<<' + str(rv) + '>>>\n\n')
-                return rv
+                cleanbuf = ''
+            if not r or len(re.findall(prompt_regex, cleanbuf)) > 0:
+                if cleanbuf:
+                    cleanbuf = cleanbuf.replace('\r', '\n')
+                self.logger.info('\n\nread complete: <<<' + str(cleanbuf) + '>>>\n\n')
+                return cleanbuf
 
-    def _ssh_command(self, context, ssh, channel, command, prompt_regex):
-        if self.fakedata:
-            print command
-            if command in self.fakedata:
-                print self.fakedata[command]
-                return self.fakedata[command]
-            else:
-                return ''
-        else:
-            self._ssh_write(context, ssh, channel, command + '\n')
-            rv = self._ssh_read(context, ssh, channel, prompt_regex)
-            if '\n%' in rv.replace('\r', '\n'):
-                es = 'CLI error message: ' + rv
-                self._log(context, es)
-                raise Exception(es)
-            return rv
+    def command(self, command, prompt_regex):
+        self._write(command + '\n')
+        rv = self._read(prompt_regex)
+        if '\n%' in rv.replace('\r', '\n'):
+            es = 'CLI error message: ' + rv
+            self.logger.info(es)
+            raise Exception(es)
+        return rv
 
-    def _connect(self, context):
-        if self.fakedata:
-            return None, None, None
-        try:
-            domain = context.reservation.domain
-        except:
-            domain = 'Global'
 
-        api = CloudShellAPISession(context.connectivity.server_address,
-                                   token_id=context.connectivity.admin_auth_token,
-                                   port=context.connectivity.cloudshell_api_port,
-                                   domain=domain)
-        address = context.resource.address
-        user = context.resource.attributes['User']
-        password = api.DecryptPassword(context.resource.attributes['Password']).Value
+def decrypt_password(context, encrypted_password):
+    """
 
-        ssh, channel, o = self._ssh_connect(context, address, 22, user, password, ' > ')
+    :param context: ResourceCommandContext | AutoLoadCommandContext
+    :param encrypted_password: str
+    :rtype: str
+    """
+    try:
+        domain = context.reservation.domain
+    except:
+        domain = 'Global'
 
-        e = self._ssh_command(context, ssh, channel, 'enable', ' # |ssword:')
-        if 'ssword:' in e:
-            enable_password = api.DecryptPassword(context.resource.attributes['Enable Password']).Value
-            self._ssh_command(context, ssh, channel, enable_password, ' # ')
-        return ssh, channel, o
+    api = CloudShellAPISession(context.connectivity.server_address,
+                               token_id=context.connectivity.admin_auth_token,
+                               port=context.connectivity.cloudshell_api_port,
+                               domain=domain)
+    return api.DecryptPassword(encrypted_password).Value
 
-    def _disconnect(self, context, ssh, channel):
-        if self.fakedata:
-            return
-        self._ssh_disconnect(context, ssh, channel)
 
+def get_logger_from_context(context):
+    """
+
+    :param context: ResourceCommandContext | AutoLoadCommandContext
+    :rtype: Logger
+    """
+    try:
+        resid = context.reservation.reservation_id
+    except:
+        resid = 'out-of-reservation'
+    try:
+        resourcename = context.resource.fullname
+    except:
+        resourcename = 'no-resource'
+    return get_qs_logger(resid, 'Mellanox-OS-L2', resourcename)
+
+
+class MellanoxOsDriver(ResourceDriverInterface):
     def __init__(self):
         """
         ctor must be without arguments, it is created with reflection at run time
         """
-        self.fakedata = None
+        self.ssh_manager = None
 
-    def initialize(self, context):
+    def initialize(self, context, ssh_manager=None):
         """
         Initialize the driver session, this function is called everytime a new instance of the driver is created
         This is a good place to load and cache the driver configuration, initiate sessions etc.
         :param InitCommandContext context: the context the command runs on
+        :param ssh_manager SSHManager: fake SSH manager for unit tests
         """
-        pass
+        self.ssh_manager = ssh_manager
+        return 'Finished initializing'
 
     # <editor-fold desc="Connectivity Provider Interface (Optional)">
 
@@ -178,9 +169,24 @@ class MellanoxOsDriver (ResourceDriverInterface):
         :rtype: str
         """
 
-        ssh, channel, _ = self._connect(context)
+        logger = get_logger_from_context(context)
+        logger.info('Request: %s' % request)
+
+        sshs = self.ssh_manager or SSHManager(logger,
+                          context.resource.address,
+                          22,
+                          get_attribute_by_name(context=context, attribute_name='User'),
+                          decrypt_password(context, get_attribute_by_name(context=context, attribute_name='Password')),
+                          ' > ')
+
+        o = sshs.command('enable', ' # |ssword:')
+        if 'ssword:' in o:
+            enable_password = decrypt_password(context,
+                                               get_attribute_by_name(context=context, attribute_name='Enable Password'))
+            sshs.command(enable_password, ' # ')
+
         try:
-            self._ssh_command(context, ssh, channel, 'configure terminal', ' # ')
+            sshs.command('configure terminal', ' # ')
             parking_vlan = '2098'
             try:
                 def add_vlan_handler(qr):
@@ -193,10 +199,10 @@ class MellanoxOsDriver (ResourceDriverInterface):
                     try:
                         addr = '/'.join(qr.actionTarget.fullAddress.split('/')[1:])
                         vlan = qr.connectionParams.vlanId
-                        self._ssh_command(context, ssh, channel, 'vlan %s' % vlan, ' # ')
-                        self._ssh_command(context, ssh, channel, 'exit', ' # ')
-                        self._ssh_command(context, ssh, channel, 'interface ethernet %s switchport mode access' % addr, ' # ')
-                        self._ssh_command(context, ssh, channel, 'interface ethernet %s switchport access vlan %s' % (addr, vlan), ' # ')
+                        sshs.command('vlan %s' % vlan, ' # ')
+                        sshs.command('exit', ' # ')
+                        sshs.command('interface ethernet %s switchport mode access' % addr, ' # ')
+                        sshs.command('interface ethernet %s switchport access vlan %s' % (addr, vlan), ' # ')
                         return ConnectivitySuccessResponse(qr, 'Success')
                     except Exception as e:
                         return ConnectivityErrorResponse(qr, str(e))
@@ -211,11 +217,11 @@ class MellanoxOsDriver (ResourceDriverInterface):
                             qr = ConnectivityActionRequest()
                         addr = '/'.join(qr.actionTarget.fullAddress.split('/')[1:])
                         vlan = qr.connectionParams.vlanId
-                        self._ssh_command(context, ssh, channel, 'interface ethernet %s switchport access vlan %s' % (addr, parking_vlan), ' # ')
+                        sshs.command('interface ethernet %s switchport access vlan %s' % (addr, parking_vlan), ' # ')
                         try:
-                            self._ssh_command(context, ssh, channel, 'no vlan %s' % vlan, ' # ')
+                            sshs.command('no vlan %s' % vlan, ' # ')
                         except:
-                            self._log(context, 'Ignoring the error from "no vlan %s"' % vlan)
+                            logger.info('Ignoring the error from "no vlan %s"' % vlan)
                         return ConnectivitySuccessResponse(qr, 'Success')
                     except Exception as e:
                         return ConnectivityErrorResponse(qr, str(e))
@@ -224,9 +230,9 @@ class MellanoxOsDriver (ResourceDriverInterface):
                                                   add_vlan_action=add_vlan_handler,
                                                   remove_vlan_action=remove_vlan_handler)
             finally:
-                self._ssh_command(context, ssh, channel, 'exit', ' # ')
+                sshs.command('exit', ' # ')
         finally:
-            self._disconnect(context, ssh, channel)
+            sshs.disconnect()
 
     # </editor-fold>
 
@@ -239,12 +245,10 @@ class MellanoxOsDriver (ResourceDriverInterface):
         :return Attribute and sub-resource information for the Shell resource you can return an AutoLoadDetails object
         :rtype: AutoLoadDetails
         """
+
         # See below some example code demonstrating how to return the resource structure
         # and attributes. In real life, of course, if the actual values are not static,
         # this code would be preceded by some SNMP/other calls to get the actual resource information
-
-        resources = []
-        attributes = []
 
 
         def MakeAutoLoadResource(model, name, relative_address, unique_identifier=None):
@@ -255,7 +259,6 @@ class MellanoxOsDriver (ResourceDriverInterface):
             rv.unique_identifier = unique_identifier
             return rv
 
-
         def MakeAutoLoadAttribute(relative_address, attribute_name, attribute_value):
             rv = AutoLoadAttribute()
             rv.relative_address = relative_address
@@ -263,9 +266,19 @@ class MellanoxOsDriver (ResourceDriverInterface):
             rv.attribute_value = attribute_value
             return rv
 
-        ssh, channel, _ = self._connect(context)
+        resources = []
+        attributes = []
 
-        show_version = self._ssh_command(context, ssh, channel, 'show version', ' # ')
+        logger = get_logger_from_context(context)
+
+        sshs = self.ssh_manager or SSHManager(logger,
+                          context.resource.address,
+                          22,
+                          get_attribute_by_name(context=context, attribute_name='User'),
+                          decrypt_password(context, get_attribute_by_name(context=context, attribute_name='Password')),
+                          ' > ')
+
+        show_version = sshs.command('show version', ' > ')
         header2attr = {
             'Version summary': 'OS Version',
             'Product model': 'Model',
@@ -278,7 +291,6 @@ class MellanoxOsDriver (ResourceDriverInterface):
                 if line.startswith(header + ':'):
                     value = line.split(header + ':')[1].strip().replace('\\"', '')
                     attributes.append(MakeAutoLoadAttribute('', header2attr[header], value))
-
 
         header2attr = {
             'HW address': 'MAC Address',
@@ -293,7 +305,7 @@ class MellanoxOsDriver (ResourceDriverInterface):
             'Description': 'Port Description',
         }
 
-        show_interfaces = self._ssh_command(context, ssh, channel, 'show interfaces', ' # ')
+        show_interfaces = sshs.command('show interfaces', ' > ')
 
         chassis = set()
         modules = set()
@@ -305,26 +317,32 @@ class MellanoxOsDriver (ResourceDriverInterface):
                 addr = '1/' + addr0
                 if '1' not in chassis:
                     chassis.add('1')
-                    resources.append(MakeAutoLoadResource(model='Generic Chassis', name='Chassis 1',    relative_address='1'))
-                resources.append(MakeAutoLoadResource(model='Generic Port',        name=addr0,          relative_address=addr))
+                    resources.append(
+                        MakeAutoLoadResource(model='Generic Chassis', name='Chassis 1', relative_address='1'))
+                resources.append(MakeAutoLoadResource(model='Generic Port', name=addr0, relative_address=addr))
             elif line.startswith('Eth'):
                 addr = line.strip().replace('Eth', '')
                 aa = addr.split('/')
                 if len(aa) == 2:
                     if aa[0] not in chassis:
                         chassis.add(aa[0])
-                        resources.append(MakeAutoLoadResource(model='Generic Chassis',     name='Chassis ' + aa[0],  relative_address=aa[0]))
-                    resources.append(MakeAutoLoadResource(model='Generic Port',            name='Port ' + aa[1],     relative_address=addr))
+                        resources.append(MakeAutoLoadResource(model='Generic Chassis', name='Chassis ' + aa[0],
+                                                              relative_address=aa[0]))
+                    resources.append(
+                        MakeAutoLoadResource(model='Generic Port', name='Port ' + aa[1], relative_address=addr))
                 elif len(aa) == 3:
                     if aa[0] not in chassis:
                         chassis.add(aa[0])
-                        resources.append(MakeAutoLoadResource(model='Generic Chassis',     name='Chassis ' + aa[0],  relative_address=aa[0]))
+                        resources.append(MakeAutoLoadResource(model='Generic Chassis', name='Chassis ' + aa[0],
+                                                              relative_address=aa[0]))
                     if aa[0] + '/' + aa[1] not in modules:
                         modules.add(aa[0] + '/' + aa[1])
-                        resources.append(MakeAutoLoadResource(model='Generic Module',      name='Module ' + aa[1],   relative_address=aa[0] + '/' + aa[1]))
-                    resources.append(MakeAutoLoadResource(model='Generic Port',            name='Port ' + aa[2],     relative_address=addr))
+                        resources.append(MakeAutoLoadResource(model='Generic Module', name='Module ' + aa[1],
+                                                              relative_address=aa[0] + '/' + aa[1]))
+                    resources.append(
+                        MakeAutoLoadResource(model='Generic Port', name='Port ' + aa[2], relative_address=addr))
                 else:
-                    self._log(context, 'Unhandled address format ' + addr)
+                    logger.error('Unhandled address format ' + addr)
             else:
                 line = line.strip()
                 for header in header2attr:
@@ -351,7 +369,7 @@ class MellanoxOsDriver (ResourceDriverInterface):
                             continue
                         attributes.append(MakeAutoLoadAttribute(addr, header2attr[header], value))
 
-        self._disconnect(context, ssh, channel)
+        sshs.disconnect()
 
         rv = AutoLoadDetails()
         rv.resources = resources
@@ -368,5 +386,3 @@ class MellanoxOsDriver (ResourceDriverInterface):
         This is a good place to close any open sessions, finish writing to log files
         """
         pass
-
-
